Move is_real_function and real_function_value from backend_interface to interface
