+++ python/tlm_adjoint_firedrake/firedrake_equations.py
@@ -22,8 +22,7 @@ from .backend import Tensor, TestFunction, TrialFunction, backend_Function, \
     backend_assemble
 from .interface import function_assign, function_comm, function_get_values, \
     function_local_size, function_new, function_set_values, function_space, \
-    is_function
-from .backend_interface import is_real_function, real_function_value
+    is_function, is_real_function, real_function_value
 from .backend_code_generator_interface import assemble, matrix_multiply
 
 from .base_equations import Equation, EquationException, NullSolver, \
